libs/voice-pipeline/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voice-pipeline"
version = "0.1.0"
description = "A modular, streaming voice conversation pipeline for building voice agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Paulo", email = "paulo@example.com" }
]
keywords = [
    "voice",
    "pipeline",
    "streaming",
    "conversation",
    "asr",
    "tts",
    "llm",
    "vad",
    "real-time",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = []

[project.optional-dependencies]
# Desenvolvimento e testes
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0",
    "black>=24.0.0",
    "mypy>=1.8.0",
]

# Memory stores persistentes
redis = [
    "redis>=4.5.0",
]
sqlite = [
    "aiosqlite>=0.19.0",
]

# Observabilidade
tracing = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
]

# Processamento de áudio
audio = [
    "numpy>=1.24.0",
    "soundfile>=0.12.0",
]

# Providers populares
openai = [
    "openai>=1.0.0",
]
anthropic = [
    "anthropic>=0.18.0",
]

# Todas as dependências opcionais
all = [
    "redis>=4.5.0",
    "aiosqlite>=0.19.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "numpy>=1.24.0",
]

[project.urls]
Homepage = "https://github.com/example/voice-pipeline"
Documentation = "https://github.com/example/voice-pipeline#readme"
Repository = "https://github.com/example/voice-pipeline"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
]

[tool.hatch.build.targets.wheel]
packages = ["src/voice_pipeline"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
